subsumed ProfileLink into MobileNav
